Add BackchannelIndustrialAppStoreHttpClient
`BackchannelIndustrialAppStoreHttpClient` is a new Industrial App Store client that is intended to perform IAS operations on behalf of a user outside of an app's HTTP request pipeline.

It differs from `IndustrialAppStoreHttpClient` by accepting an `ITokenStore` on every operation instead of an `HttpContext`, which is used to retrieve the access token for outgoing requests.