Nullable reference types + session ID generator
Add `SessionIdGenerator` property to `IndustrialAppStoreAuthenticationOptions`. Purpose of this callback is to allow an app to define how a unique identifier for a given authentication session is generated.

Default behaviour is still to create a GUID for each session. However, an app may wish to generate its own identifiers based on e.g. a fingerprint for the user/device/browser combination.