Add support for defining a Content-Security-Policy header
Moves custom headers services and middleware from Authentication to a new AspNetCore project.

Adds new services and middleware for defining a Content Security Policy via configuration.

The middleware allows multiple cascading CSPs to be defined and assigned priorities. A composite CSP is built for each outgoing response and assigned to the `Content-Security-Policy` header.

Also updates templates to use a CSP by default and adds documentation to the templated app's readme file to tell the developer where to go to learn about customising it.