Use access token provider in dependency injection scenarios (#86)
* Add AccessTokenProvider

AccessTokenProvider is a new scoped service that defines a factory property for retrieving an access token for the current scope.

* Use AccessTokenProvider to authenticate requests

DefaultIndustrialAppStoreHttpFactory now creates HttpMessageHandler instances that authenticate outgoing requests using the AccessTokenProvider service.

* Add extensions for configuring AccessTokenProvider

IndustrialAppStoreExtensions.AddIndustrialAppStoreApiServices now adds a default AccessTokenProvider service that never returns an access token.

New AddAccessTokenProvider extension methods for  IIndustrialAppStoreBuilder have been added to allow the default AccessTokenProvider registration to be replaced with a registration that returns an access token for the current scope.

* Update READMEs

* Remove custom DelegatingHandler

Removes the custom CLI DelegatingHandler for API authentication. Authentication will be provided via the AccessTokenProvider service instead.

* Register CLI-specific AccessTokenProvider service

Overrides the default AccessTokenProvider registration with one that uses the IndustrialAppStoreSessionManager service to provide the access token for outgoing requests.

* Remove custom HTTP factory and auth handler

Removes the custom TokenStoreHttpFactory and TokenStoreAuthenticationHandler classes. The functionality from these types will be replaced with a custom AccessTokenProvider registration that uses the ITokenStore service to retrieve the access token for an outgoing request.

* Use ITokenStore with AccessTokenProvider

Replaces the default AccessTokenProvider registration with one that uses the ITokenStore for the current scope to retrieve access tokens.