HTTP factory improvements (#88)
* Remove CreateHandler

Removes the CreateHandler method from the IIndustrialAppStoreHttpFactory interface. This method can be an internal implementation detail of the IndustrialAppStoreHttpFactory base class.

* Update IndustrialAppStoreHttpFactory

Moves code for setting e.g. default HTTP version for an HttpClient to a new virtual ConfigureHttpClient method.

Moves IHttpClientFactory/IHttpMessageHandlerFactory client name for IndustrialAppStoreHttpClient to an internal const.

* Update doc comments

* Add static access token provider extension

Adds an AddStaticAccessTokenProvider extension method for IIndustrialAppStoreBuilder that allows a fixed access token to be used. This is primarily intended for testing scenarios but may have some use in e.g. LINQPad scripts.

* Update PACKAGE_README.md