Mitigation for CVE-2025-55247 (#111)
* Update package references

* Add top-level Microsoft.Build reference

Adds a top-level reference to Microsoft.Build to two sample projects affected by CVE-2025-55247 (https://github.com/dotnet/announcements/issues/370).

The advisory states that projects are vulnerable if they reference a vulnerable version, even if the latest SDK is installed. Microsoft.Build is a transitive reference that we inherit from Microsoft.VisualStudio.Web.CodeGeneration.Design. At present, there is not a version of this package available that mitigates the CVE, hence the use of a top-level reference to mitigate it.

* Update version.json